+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+
+## 2026.04.24
+
+### Added
+
+- Designed abstract class `ArtificialNeuronSecondary` for the
+  ArtificialNeuron component
+- Implemented secondary methods (`scaleWeights`, `sumWeights`,
+  `dotProduct`) using only kernel and Standard methods
+- Implemented `toString`, `equals`, and `hashCode` using only kernel
+  methods
+
+
 ## 2026.04.24
 
 ### Added